requirements.txt
tensorflow-cpu>=2.13,<2.16
opencv-python-headless<4.10
numpy>=1.23.5,<2.0.0
Pillow
pillow-heif
streamlit==1.54.0
